requirements.streamlit.txt
# ============================================
# Streamlit Cloud Deployment Requirements
# ============================================

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# API & HTTP
requests>=2.31.0
httpx>=0.24.0

# Database
neo4j>=5.12.0

# OpenAI & LLM
openai>=1.3.0

# GraphRAG Core
nano-graphrag>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.7.0

# MCP Integration
mcp>=0.9.0
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20

# Search
duckduckgo-search>=3.9.0

# Monitoring
psutil>=5.9.0

# Optional: Visualization
# pyvis>=0.3.2
# networkx>=3.1
